Move CompactCreditPoolCards into WorkspaceCreditPoolCards.tsx
Keeps the credit-pool card component colocated with the other
pool-card building blocks it composes.
